requirements.txt
Flask>=3.0,<3.2
Flask-RESTful>=0.3.10,<0.4
Flask-SQLAlchemy>=3.1,<3.2
Flask-Migrate>=4.0,<4.1
SQLAlchemy>=2.0,<2.1
Alembic>=1.13,<1.14
python-dotenv>=1.0,<2.0
psycopg[binary]>=3.1,<3.2
passlib[bcrypt]>=1.7,<1.8
requests>=2.31,<2.33